requirements.txt
google-generativeai
python-dotenv
colorlog
rich
absl-py